fixed duplicated task when showing a specific task in a modal
